features: Connect.feature: Update default config in firmware test
The default configuration for the application has changes. Update the
test with the new config values.

The firmware test checks that the device reports its default
configuration after an established connection. If there is a mismatch
between the expected default values and the reported default values
the firmware test fails.

Active wait timeout (actwt) and movement resolution (mvres) are now set
to 120. This is to avoid the possibility of GPS data getting requested
during a GPS search because the timeouts are too close to the
GPS timeout. It doesnt hurt to give the application some wiggle-room.

The accelerometer threshold (acct) is now set to 10 m/s2 instead of 100.
100 m/s2 is simply an unrealistic threshold and exceeds the
measuring range of the ADXL362, present on the Thingy91.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>